requirements.txt
pandas
numpy
matplotlib
plotly.express
ipykernel